+++ demisto_sdk/__main__.py
@@ -3413,6 +3413,12 @@ def update_content_graph(
     help="The demisto-sdk ref to use for the pre-commit hooks",
     default="",
 )
+@click.option(
+    "--dry-run",
+    help="Whether to run the pre-commit hooks in dry-run mode, which will only create the config file",
+    is_flag=True,
+    default=False,
+)
 @click.argument(
     "file_paths",
     nargs=-1,
@@ -3435,6 +3441,7 @@ def pre_commit(
     show_diff_on_failure: bool,
     sdk_ref: str,
     file_paths: Iterable[Path],
+    dry_run: bool,
     **kwargs,
 ):
     from demisto_sdk.commands.pre_commit.pre_commit_command import pre_commit_manager
@@ -3445,7 +3452,7 @@ def pre_commit(
         )
     input_files = []
     if input:
-        input_files = [Path(i) for i in input.split(",")]
+        input_files = [Path(i) for i in input.split(",") if i]
     elif file_paths:
         input_files = list(file_paths)
     if skip:
@@ -3465,6 +3472,7 @@ def pre_commit(
             verbose,
             show_diff_on_failure,
             sdk_ref=sdk_ref,
+            dry_run=dry_run,
         )
     )